refactor(api): move task endpoints into dedicated hueyos router
